% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{find_one_lambda}
\alias{find_one_lambda}
\title{Implements ridge regression with many predictors for one value of lambda (a helper function)}
\usage{
find_one_lambda(dat, response, lambda)
}
\arguments{
\item{dat}{A data frame}

\item{response}{The name of a response variable in the data frame (unquoted)}

\item{lambda}{A singular value to try as a penalty term}
}
\value{
A data frame of coefficients
}
\description{
This function computes coefficients for ridge regression with one lambda
All columns of the provided data frame are used as predictors, except the
one specified as a response.
}
\details{
No interaction terms are included.
}
