% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{get_betas}
\alias{get_betas}
\title{Computes ridge regression on the training and testing data sets}
\usage{
get_betas(dat, response, lambda)
}
\arguments{
\item{dat}{the data frame}

\item{response}{The name of a response variable in the data frame (unquoted)}

\item{lambda}{A penalty term to try}
}
\value{
data frame of coeffecients
}
\description{
Returns coeffecients
}
\details{
No interaction terms are included.
}
