% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{get_predictions}
\alias{get_predictions}
\title{Get predictions from the beta values}
\usage{
get_predictions(dat, response, ridge_results)
}
\arguments{
\item{dat}{the data frame to be turned into design matrix}

\item{response}{the response variable}

\item{ridge_results}{the beta values}
}
\value{
matrix y with predicted values
}
\description{
Returns
}
\details{
No interaction terms are included.
}
