% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{get_test_error}
\alias{get_test_error}
\title{Returns the mean squared error value for that lambda}
\usage{
get_test_error(y_train, y_test)
}
\arguments{
\item{y_train}{A data frame to test the model on}

\item{y_test}{A data frame to construct the model from}
}
\value{
mean squared error
}
\description{
No interaction terms are included.
}
