% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_regression.R
\name{lasso_gd}
\alias{lasso_gd}
\title{Implements lasso linear regression by gradient descent}
\usage{
lasso_gd(
  dat,
  response,
  learning_rate = 0.001,
  lambda = 0.5,
  iterations = 50000
)
}
\arguments{
\item{dat}{A data frame}

\item{response}{The name of a response variable in the data frame (unquoted)}

\item{learning_rate}{The learning rate controls how quickly the model is adapted to the problem (default 0.05)}

\item{lambda}{Tuning Parameter that controls the bias-variance trade off (default 0.8)}

\item{iterations}{Number of iterations for gradient descent to run (default 50000)}
}
\value{
A data frame of coefficients
}
\description{
This function computes coefficients for multiple regression by lasso l1 penalization and gradient descent
All columns of the provided data frame are used as predictors, except the
one specified as a response.
}
