% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_regression.R
\name{lasso_loss}
\alias{lasso_loss}
\title{Computes derivate of MSE loss function with additional l1 regularization term}
\usage{
lasso_loss(num_explanatory, x, y, y_pred, lambda, thetas)
}
\arguments{
\item{num_explanatory}{Number of explanatory variables in given data set}

\item{x}{Values of explanatory variables}

\item{y}{Values of response variable from data set}

\item{y_pred}{Fitted values predicted from current thetas}

\item{lambda}{Tuning Parameter that controls the bias-variance trade off}

\item{thetas}{Coefficients/Weights of multivariate regression}
}
\value{
A vector of the derivatives of the MSE loss function
}
\description{
Computes derivate of MSE loss function with additional l1 regularization term
}
