% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_alt.R
\name{loss_function}
\alias{loss_function}
\title{Loss/Cost Function for multivariate gradient descent}
\usage{
loss_function(X, y, m, theta)
}
\description{
Loss/Cost Function for multivariate gradient descent
}
