% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_viz.R
\name{residual_calculate}
\alias{residual_calculate}
\title{Creates a data frame with residuals}
\usage{
residual_calculate(data, response, prediction)
}
\arguments{
\item{response}{The name of a response variable in the data frame (unquoted)}

\item{prediction}{A data frame with the predictions}

\item{dat}{A data frame}
}
\value{
A data frame with the residuals
}
\description{
Creates a data frame with residuals
}
