% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_viz.R
\name{residual_viz}
\alias{residual_viz}
\title{Creates a histogram displaying the residuals}
\usage{
residual_viz(data, response, prediction)
}
\arguments{
\item{response}{The name of a response variable in the data frame (unquoted)}

\item{prediction}{A data frame with the predictions}

\item{dat}{A data frame}
}
\value{
A histogram with the residuals
}
\description{
Creates a histogram displaying the residuals
}
